requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
supabase
cachetools
python-dotenv
sqlmodel
httpx
pytest
requests
starlette
python-multipart
gunicorn
